Fold KeywordSearch into SpokenSearch.Keyword; Query/ is now Visual + Spoken
