--- Sources/PalmierPro/Search/Query/KeywordSearch.swift
@@ -1,30 +1,0 @@
-import Foundation
-
-/// Exact-keyword tier of SpokenSearch: cached transcripts, all query words in any order.
-enum KeywordSearch {
-    /// Query split into words, edge punctuation stripped (so "budget," → "budget").
-    static func terms(in query: String) -> [String] {
-        query.split(whereSeparator: \.isWhitespace)
-            .map { $0.trimmingCharacters(in: .punctuationCharacters) }
-            .filter { !$0.isEmpty }
-    }
-
-    static func matches(_ text: String, terms: [String]) -> Bool {
-        terms.allSatisfy { text.range(of: $0, options: [.caseInsensitive, .diacriticInsensitive]) != nil }
-    }
-
-    static func search(query: String, assets: [(id: String, url: URL)], limit: Int = 20) -> [SpokenSearch.Hit] {
-        let terms = terms(in: query)
-        guard !terms.isEmpty else { return [] }
-
-        var hits: [SpokenSearch.Hit] = []
-        for asset in assets {
-            guard let transcript = TranscriptCache.cachedOnDisk(for: asset.url) else { continue }
-            for segment in transcript.segments where matches(segment.text, terms: terms) {
-                hits.append(SpokenSearch.Hit(assetID: asset.id, start: segment.start, end: segment.end, text: segment.text))
-                if hits.count >= limit { return hits }
-            }
-        }
-        return hits
-    }
-}
